Moved things around so that it's not cluttering the main file
